Moved SelectableInstance into its own folder
